requirements.txt
flask==3.0.0
apscheduler==3.10.4
psutil==5.9.7
pyyaml==6.0.1
requests==2.31.0

# For building standalone collector executable
pyinstaller==6.3.0
